requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
pyyaml
